samples/02-resilient-connections/python/requirements.txt
pyodbc
tenacity
python-dotenv